move new functions into TypeCheck.External
